// src/lib/Damage/DamageCalculation/enums.ts
export const enum ContainerTypes {
  Normal = 'normal',
  Options = 'options',
}

export const enum CalculationContainerIds {
  DamageType = 'damage_type',
  Base = 'base',
  BaseAtk = 'base/atk',
  BaseMatk = 'base/matk',
  BaseDualSword = 'base/dual_sword',
  BaseTwoHanded = 'base/two_handed',
  TargetResistance = 'target_resistance',
  LevelDifference = 'level_difference',
  TargetDefBase = 'target_def_base',
  Pierce = 'pierce',
  SkillConstant = 'skill/constant',
  UnsheatheAttackConstant = 'unsheathe_attack/constant',
  OtherConstant = 'other_constant',
  SkillMultiplier = 'skill/multiplier',
  // Critical = 'critical',
  CriticalDamage = 'critical/critical_damage',
  CriticalRate = 'critical/critical_rate',
  Accuracy = 'accuracy',
  Stability = 'stability',
  Critical_Accuracy_Stability = 'critical|accuracy|stability',
  RangeDamage = 'range_damage',
  UnsheatheAttackMultiplier = 'unsheathe_attack/multiplier',
  StrongerAgainstElement = 'stronger_against_element',
  Proration = 'proration',
  ComboMultiplier = 'combo_multiplier',
  // SkillLongRange = 'skill/long_range',
  OtherMultiplier = 'other_multiplier',
}

export const enum CalculationItemIds {
  Physical = 'physical',
  Magic = 'magic',
  Atk = 'atk',
  Matk = 'matk',
  AtkRate = 'atk_rate',
  MatkRate = 'matk_rate',
  SubAtk = 'sub_atk',
  SubStability = 'sub_stability',
  SkillLevelTwoHanded = 'skill_level_two_handed',
  TargetPhysicalResistance = 'target_physical_resistance',
  TargetMagicResistance = 'target_magic_resistance',
  CharacterLevel = 'character_level',
  TargetLevel = 'target_level',
  TargetDef = 'target_def',
  TargetMdef = 'target_mdef',
  PhysicalPierce = 'physical_pierce',
  MagicPierce = 'magic_pierce',
  SkillConstant = 'skill_constant',
  UnsheatheAttackConstant = 'unsheathe_attack_constant',
  OtherConstant = 'other_constant',
  SkillMultiplier = 'skill_multiplier',
  CriticalDamage = 'critical_damage',
  MagicCriticalDamageConversionRate = 'magic_critical_damage_conversion_rate',
  CriticalRate = 'critical_rate',
  MagicCriticalRateConversionRate = 'magic_critical_rate_conversion_rate',
  TargetCriticalRateResistance = 'target_critical_rate_resistance',
  TargetCriticalRateResistanceTotal = 'target_critical_rate_resistance_total',
  ShortRangeDamage = 'short_range_damage',
  LongRangeDamage = 'long_range_damage',
  UnsheatheAttackMultiplier = 'unsheathe_attack_multiplier',
  StrongerAgainstNeutral = 'stronger_against_neutral',
  StrongerAgainstFire = 'stronger_against_fire',
  StrongerAgainstWater = 'stronger_against_water',
  StrongerAgainstEarth = 'stronger_against_earth',
  StrongerAgainstWind = 'stronger_against_wind',
  StrongerAgainstLight = 'stronger_against_light',
  StrongerAgainstDark = 'stronger_against_dark',
  Proration = 'proration',
  ComboMultiplier = 'combo_multiplier',
  // SkillLevelLongRange = 'skill_level_long_range',
  Stability = 'stability',
  Accuracy = 'accuracy',
  SkillRealMpCost = 'skill_real_mp_cost',
  TargetDodge = 'target_dodge',
  PromisedAccuracyRate = 'promised_accuracy_rate',
  // ProbabilityOfGraze = 'probability_of_graze',
  OtherMultiplier = 'other_multiplier',
}